Refactor test fixtures and utility functions
