+++ tests/integration/local/syft_worker_deletion_test.py
@@ -1,9 +1,5 @@
 # stdlib
-from collections.abc import Generator
-from collections.abc import Iterable
-from itertools import product
 import operator
-from secrets import token_hex
 import time
 from typing import Any
 
@@ -13,65 +9,22 @@ import pytest
 
 # syft absolute
 import syft as sy
-from syft.orchestra import ServerHandle
+from syft.orchestra import ClientAlias
 from syft.service.job.job_stash import JobStatus
 from syft.service.response import SyftError
 
+# relative
+from .conftest import matrix
+
 # equivalent to adding this mark to every test in this file
 pytestmark = pytest.mark.local_server
-
-
-@pytest.fixture()
-def server_args() -> dict[str, Any]:
-    return {}
-
-
-@pytest.fixture
-def server(server_args: dict[str, Any]) -> Generator[ServerHandle, None, None]:
-    _server = sy.orchestra.launch(
-        **{
-            "name": token_hex(8),
-            "dev_mode": True,
-            "reset": True,
-            "n_consumers": 3,
-            "create_producer": True,
-            "queue_port": None,
-            "local_db": False,
-            **server_args,
-        }
-    )
-    # startup code here
-    yield _server
-    # Cleanup code
-    _server.python_server.cleanup()
-    _server.land()
-
-
-def matrix(
-    *,
-    excludes_: Iterable[dict[str, Any]] | None = None,
-    **kwargs: Iterable,
-) -> list[dict[str, Any]]:
-    args = ([(k, v) for v in vs] for k, vs in kwargs.items())
-    args = product(*args)
-
-    if excludes_ is None:
-        excludes_ = []
-    excludes_ = [kv.items() for kv in excludes_]
-
-    args = (
-        arg
-        for arg in args
-        if not any(all(kv in arg for kv in kvs) for kvs in excludes_)
-    )
-
-    return [dict(kvs) for kvs in args]
 
 
 SERVER_ARGS_TEST_CASES = {
     "n_consumers": [1],
     "dev_mode": [True, False],
     "thread_workers": [True, False],
+    "create_producer": [True],
 }
 
 
@@ -90,10 +43,8 @@ class FlakyMark(RuntimeError):
 )
 @pytest.mark.parametrize("force", [True, False])
 def test_delete_idle_worker(
-    server: ServerHandle, force: bool, server_args: dict[str, Any]
+    client: ClientAlias, force: bool, server_args: dict[str, Any]
 ) -> None:
-    client = server.login(email="info@openmined.org", password="changethis")
-
     original_workers = client.worker.get_all()
     worker_to_delete = max(original_workers, key=operator.attrgetter("name"))
 
@@ -126,9 +77,7 @@ def test_delete_idle_worker(
 
 @pytest.mark.parametrize("server_args", matrix(**SERVER_ARGS_TEST_CASES))
 @pytest.mark.parametrize("force", [True, False])
-def test_delete_worker(server: ServerHandle, force: bool) -> None:
-    client = server.login(email="info@openmined.org", password="changethis")
-
+def test_delete_worker(client: ClientAlias, force: bool) -> None:
     data = np.array([1, 2, 3])
     data_action_obj = sy.ActionObject.from_obj(data)
     data_pointer = data_action_obj.send(client)
